backend/requirements.txt
flask==3.1.0
python-dotenv==0.21.0
fauna==2.4.0
# flask-sse==1.0.0
# redis==5.2.1
